pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vagent"
version = "0.1.0"
description = "Voice Agent using LiveKit Agents SDK"
requires-python = ">=3.10"
dependencies = [
    "livekit-agents[silero,cartesia]==1.3.6",
    "livekit-plugins-openai==1.3.6",
    "livekit-api==1.1.0",
    "faster-whisper==1.2.1",
    "ctranslate2==4.6.2",
    "nvidia-cudnn-cu12",
    "httpx==0.28.1",
    "python-dotenv==1.2.1",
    "scipy==1.16.3",
    "numpy==2.3.5",
    "openai",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "black",
    "ruff",
]

[project.scripts]
vagent = "vagent.agent:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"*" = ["*.json"]

[tool.black]
line-length = 120
target-version = ["py310"]

[tool.ruff]
line-length = 120
target-version = "py310"
